Refactor code, move composables to a the package 'ui'
