+++ app/src/androidTest/java/com/github/mateo762/myapplication/HabitsActivityFragmentNavigationTest.kt
@@ -8,7 +8,7 @@ import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.mateo762.myapplication.habits.HabitsActivity
-import com.github.mateo762.myapplication.habits.fragments.DevelopFragment
+import com.github.mateo762.myapplication.habits.fragments.CreateHabitFragment
 import com.github.mateo762.myapplication.habits.fragments.WeekFragment
 import junit.framework.TestCase
 import org.junit.After
@@ -56,7 +56,7 @@ class HabitsActivityFragmentNavigationTest {
         // Get a reference to the current fragment
         val fragment = getCurrentFragment()
         // Check if the current fragment is a DevelopFragment
-        TestCase.assertTrue(fragment is DevelopFragment)
+        TestCase.assertTrue(fragment is CreateHabitFragment)
     }
 
     private fun getCurrentFragment(): Fragment? {
