+++ app/src/androidTest/java/com/github/mateo762/myapplication/RegisterActivityTest.kt
@@ -10,8 +10,8 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.mateo762.myapplication.login.LoginActivity
-import com.github.mateo762.myapplication.register.RegisterActivity
+import com.github.mateo762.myapplication.authentication.LoginActivity
+import com.github.mateo762.myapplication.authentication.RegisterActivity
 import org.hamcrest.Matchers
 import org.junit.After
 import org.junit.Before
