+++ app/src/main/java/com/github/mateo762/myapplication/habits/HabitsActivity.kt
@@ -5,7 +5,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.ListFragment
 import com.github.mateo762.myapplication.BaseActivity
 import com.github.mateo762.myapplication.R
-import com.github.mateo762.myapplication.habits.fragments.DevelopFragment
+import com.github.mateo762.myapplication.habits.fragments.CreateHabitFragment
 import com.github.mateo762.myapplication.habits.fragments.WeekFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
@@ -25,7 +25,7 @@ class HabitsActivity : BaseActivity() {
         lateinit var selectedFragment: Fragment
         when (it.itemId) {
             R.id.developFragment -> {
-                selectedFragment = DevelopFragment()
+                selectedFragment = CreateHabitFragment()
             }
             R.id.weekFragment -> {
                 selectedFragment = WeekFragment()
