+++ app/src/main/java/com/github/mateo762/myapplication/habits/fragments/CreateHabitFragment.kt
@@ -1,0 +1,30 @@
+package com.github.mateo762.myapplication.habits.fragments
+
+import android.os.Build
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.RequiresApi
+import androidx.compose.foundation.layout.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.platform.ComposeView
+import androidx.fragment.app.Fragment
+import com.github.mateo762.myapplication.ui.habits.CreateHabitScreen
+
+
+class CreateHabitFragment : Fragment() {
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+
+        return ComposeView(requireContext()).apply {
+            setContent {
+                CreateHabitScreen()
+            }
+        }
+    }
+}
