+++ app/src/main/java/com/github/mateo762/myapplication/home/fragments/SummaryFragment.kt
@@ -1,8 +1,6 @@
 package com.github.mateo762.myapplication.home.fragments
 
-import android.annotation.SuppressLint
 import android.content.ContentValues.TAG
-import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import android.util.Log
@@ -11,22 +9,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
 import androidx.fragment.app.Fragment
-import com.github.mateo762.myapplication.R
-import com.github.mateo762.myapplication.habits.HabitsActivity
-import java.time.DayOfWeek
-import java.time.LocalTime
+import com.github.mateo762.myapplication.getHardCodedHabits
+import com.github.mateo762.myapplication.ui.home.HabitListScreen
 
 
 class SummaryFragment : Fragment() {
@@ -45,122 +32,5 @@ class SummaryFragment : Fragment() {
             throw e
         }
     }
-
-    data class Habit(
-        val name: String,
-        val daysOfWeek: List<DayOfWeek>,
-        val startTime: LocalTime,
-        val endTime: LocalTime
-    )
-
-    @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
-    @Composable
-    fun HabitListScreen(habits: List<Habit>) {
-        val context = LocalContext.current
-        Scaffold(
-            floatingActionButton = {
-                FloatingActionButton(
-                    onClick = {
-                        val intent = Intent(context, HabitsActivity::class.java)
-                        context.startActivity(intent)
-                    },
-                    backgroundColor = Color.White,
-                    contentColor = Color.Black
-                ) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.ic_new),
-                        contentDescription = "Add"
-                    )
-                }
-            },
-            modifier = Modifier.testTag("btn_new")
-        ) {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(16.dp)
-            ) {
-                items(habits) { habit ->
-                    HabitListItem(habit = habit)
-                }
-            }
-        }
-    }
-
-    @Composable
-    fun HabitListItem(habit: Habit) {
-        Card(
-            elevation = 4.dp,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 8.dp)
-                .testTag(habit.name)
-        ) {
-            Column(
-                modifier = Modifier.padding(16.dp)
-            ) {
-                Text(
-                    text = habit.name,
-                    style = MaterialTheme.typography.h6
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = habit.daysOfWeek.joinToString(),
-                    style = MaterialTheme.typography.body1
-                )
-                Spacer(modifier = Modifier.height(4.dp))
-                Text(
-                    text = "Start time: ${habit.startTime}",
-                    style = MaterialTheme.typography.body1
-                )
-                Spacer(modifier = Modifier.height(4.dp))
-                Text(
-                    text = "End time: ${habit.endTime}",
-                    style = MaterialTheme.typography.body1
-                )
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun getHardCodedHabits(): List<Habit> {
-        return listOf(
-            Habit(
-                name = "Morning Walk",
-                daysOfWeek = listOf(DayOfWeek.MONDAY, DayOfWeek.WEDNESDAY, DayOfWeek.FRIDAY),
-                startTime = LocalTime.of(6, 30),
-                endTime = LocalTime.of(7, 30),
-            ),
-            Habit(
-                name = "Reading",
-                daysOfWeek = listOf(DayOfWeek.TUESDAY, DayOfWeek.THURSDAY, DayOfWeek.SATURDAY),
-                startTime = LocalTime.of(20, 0),
-                endTime = LocalTime.of(21, 0)
-            ),
-            Habit(
-                name = "Meditation",
-                daysOfWeek = listOf(DayOfWeek.SUNDAY),
-                startTime = LocalTime.of(7, 0),
-                endTime = LocalTime.of(8, 0)
-            ),
-            Habit(
-                name = "Walking",
-                daysOfWeek = listOf(DayOfWeek.MONDAY, DayOfWeek.WEDNESDAY, DayOfWeek.FRIDAY),
-                startTime = LocalTime.of(6, 30),
-                endTime = LocalTime.of(7, 30)
-            ),
-            Habit(
-                name = "Gym",
-                daysOfWeek = listOf(DayOfWeek.TUESDAY, DayOfWeek.THURSDAY, DayOfWeek.SATURDAY),
-                startTime = LocalTime.of(20, 0),
-                endTime = LocalTime.of(21, 0)
-            ),
-            Habit(
-                name = "Swimming",
-                daysOfWeek = listOf(DayOfWeek.SUNDAY),
-                startTime = LocalTime.of(7, 0),
-                endTime = LocalTime.of(8, 0)
-            )
-        )
-    }
 }
+
